requirements.txt
asgiref==3.7.2
Django==4.2.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
sqlparse==0.4.4
